requirements.txt
altgraph==0.17.2
arabic-reshaper==2.1.3
arrow==0.17.0
attrs==20.3.0
base58==2.0.1
bipwallet==1.0.2
bitarray==1.2.2
click==6.6
cytoolz==0.11.0
Django==2.2
django-qr-code==2.0.1
docker-py==1.8.0
ecdsa==0.16.1
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
flake8==3.8.4
funcsigs==1.0.2
future==0.18.2
hexbytes==0.2.2
html5lib==1.1
importlib-metadata==3.3.0
iniconfig==1.1.1
jsonrpcclient==2.0.1
jsonrpcserver==3.1.1
jsonschema==3.2.0
mccabe==0.6.1
mnemonic==0.13
packaging==20.8
parsimonious==0.8.1
path==15.0.1
path.py==12.5.0
pbkdf2==1.3
pdf2image==1.16.0
pexpect==4.8.0
Pillow==8.3.2
pkg-resources==0.0.0
pluggy==0.13.1
protobuf==3.0.0a3
ptyprocess==0.6.0
py==1.10.0
pyaes==1.6.1
pycodestyle==2.6.0
pycryptodome==3.9.9
pyflakes==2.2.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
pytest==6.2.1
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
reportlab==3.6.1
requests==2.11.1
rlp==2.0.1
segno==1.3.1
sha256==0.1
six==1.15.0
sqlparse==0.4.1
tabulate==0.8.7
toml==0.10.2
toolz==0.11.1
two1==3.10.9
typing-extensions==3.7.4.3
uuid==1.30
webencodings==0.5.1
websocket-client==0.57.0
xhtml2pdf==0.2.5
zipp==3.4.0
